Reorganize mkdir_p fsutil
