+++ yatsm/io/__init__.py
@@ -10,11 +10,9 @@ Contents:
       filesystem operations
 """
 from ._api import get_readers, read_and_preprocess
-from ._util import mkdir_p
 
 
 __all__ = [
-    'mkdir_p',
     'get_readers',
     'read_and_preproces'
 ]
