+++ yatsm/utils.py
@@ -2,6 +2,7 @@
 """
 from __future__ import division
 
+import errno
 import fnmatch
 from functools import wraps
 import logging
@@ -93,6 +94,26 @@ def find(location, pattern, regex=False):
                 files.append(os.path.join(root, filename))
 
     return sorted(files)
+
+
+def mkdir_p(d):
+    """ Make a directory, ignoring error if it exists (i.e., ``mkdir -p``)
+
+    Args:
+        d (str): directory path to create
+
+    Raises:
+        OSError: Raise OSError if cannot create directory for reasons other
+        than it existing already (errno 13 "EEXIST")
+    """
+    try:
+        os.makedirs(d)
+    except OSError as err:
+        # File exists
+        if err.errno == errno.EEXIST:
+            pass
+        else:
+            raise err
 
 
 # MISC UTILITIES
